:bug: move image-path-utils to frontend where it belongs
