+++ example-29-fullstack-typescript/frontend/src/components/GalleryCard.jsx
@@ -1,7 +1,7 @@
 import { Card, CardActionArea, CardContent, CardMedia, Typography } from "@mui/material";
 import { Link } from "react-router-dom";
 import { getPlaintextSummary } from "../util/article-helpers";
-import { getImagePath } from "../../../backend/src/util/image-path-utils";
+import { getImagePath } from "../util/image-path-utils";
 
 /**
  * Displays Material UI Card containing info about the given article.
