+++ example-29-fullstack-typescript/frontend/src/util/image-path-utils.ts
@@ -1,4 +1,3 @@
-import env from "./env";
 import { joinPaths } from "./path-utils";
 
 // !!! TypeScript !!!
