+++ example-29-fullstack-typescript/frontend/src/util/path-utils.ts
